Add files
This script, when modified with a custom set of URLs, can be used as a template for scraping .txt files (or any other filetype) from any website (such as Project Gutenberg). Future potential additions include more advanced header parsing with many branches so all files of a type can be scraped (not necessary for my current purposes).